minor fixes

+++ framework/net/gw_proxy_site.class.php
@@ -20,7 +20,7 @@ class GW_Proxy_Site {
 	static function redirect($mirrorhost) 
 	{
 
-		$cnt=file_get_contents($f=GW::s('DIR/TEMP').'botcount'.date('Y-m-d'));
+		$cnt=(int)file_get_contents($f=GW::s('DIR/TEMP').'botcount'.date('Y-m-d'));
 		file_put_contents($f, $cnt+1);
 		
 		/* Set it true for debugging. */
